requirements.txt
autogen-agentchat>=0.5.0
autogen-ext[openai]>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0